HAWKULAR-1336 - configure build to fix issue with dev profile (#241)
If the "dev" profile is activated, generation of the hawkular services
dist zip is skipped to speed up the build.  The itest and docker-dist
modules are dependent on this zip, so the build will fail if the zip
is not in the local maven repo.  This change causes the itest and
docker-dist modules to be skipped completely if the "dev" profile is active.